+++ examples/JevSandbox/Source/JevSandbox/JevSandbox.Build.cs
@@ -5,6 +5,6 @@ public class JevSandbox : ModuleRules
     public JevSandbox(ReadOnlyTargetRules Target) : base(Target)
     {
         PCHUsage = PCHUsageMode.UseExplicitOrSharedPCHs;
-        PublicDependencyModuleNames.AddRange(new[] { "Core", "CoreUObject", "Engine" });
+        PublicDependencyModuleNames.AddRange(new[] { "Core", "CoreUObject", "Engine", "FunctionalTesting", "NavigationSystem" });
     }
 }